% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fmx.R
\name{as.fmx.Normal}
\alias{as.fmx.Normal}
\title{Convert \code{Normal} fit from \CRANpkg{mixsmsn} to \linkS4class{fmx}}
\usage{
\method{as.fmx}{Normal}(x, data = double(), ...)
}
\arguments{
\item{x}{\code{'Normal'} object,
returned from \link[mixsmsn]{smsn.mix} with parameter
\code{family = 'Normal'}}

\item{data}{\link[base]{numeric} \link[base]{vector}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=as.fmx.Normal]{as.fmx.Normal()}} returns an \linkS4class{fmx} object.
}
\description{
To convert \code{Normal} object (from package \CRANpkg{mixsmsn})
to \linkS4class{fmx} class.
}
\note{
\link[mixsmsn]{smsn.mix} does not offer a parameter to keep the input data, as of 2021-10-06.
}
\keyword{internal}
