% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fmx.R
\name{as.fmx.t}
\alias{as.fmx.t}
\title{Convert \code{t} fit from \CRANpkg{mixsmsn} to \link[fmx]{fmx-class}}
\usage{
\method{as.fmx}{t}(x, data = double(), ...)
}
\arguments{
\item{x}{\code{'t'} object,
returned from \link[mixsmsn]{smsn.mix} with parameter
\code{family = 't'}}

\item{data}{\link[base]{numeric} \link[base]{vector}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=as.fmx.t]{as.fmx.t()}} has not been completed yet
}
\description{
To convert \code{t} object (from package \CRANpkg{mixsmsn})
to \link[fmx]{fmx-class} class.
}
\note{
\link[mixsmsn]{smsn.mix} does not offer a parameter to keep the input data, as of 2021-10-06.
}
\keyword{internal}
