% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort_}
\alias{sort_}
\alias{sort.Skew.normal}
\alias{sort.Normal}
\alias{sort.Skew.t}
\alias{sort.t}
\title{Sort Objects from \CRANpkg{mixsmsn} by Location Parameters}
\usage{
\method{sort}{Skew.normal}(x, decreasing = FALSE, ...)

\method{sort}{Normal}(x, decreasing = FALSE, ...)

\method{sort}{Skew.t}(x, decreasing = FALSE, ...)

\method{sort}{t}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{\code{'Normal'}, \code{'Skew.normal'}, \code{'Skew.t'} object}

\item{decreasing}{\link[base]{logical} scalar. Should the sort the location parameter
be increasing (\code{FALSE}, default) or decreasing (\code{TRUE})?}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=sort.Normal]{sort.Normal()}} returns a \code{'Normal'} object.

Function \code{\link[=sort.Skew.normal]{sort.Skew.normal()}} returns a \code{'Skew.normal'} object.

Function \code{\link[=sort.Skew.t]{sort.Skew.t()}} returns a \code{'Skew.t'} object.
}
\description{
To sort an object returned from package \CRANpkg{mixsmsn} by its location parameters
}
\details{
\link[mixsmsn]{smsn.mix} does \emph{not} order the location parameter
}
\seealso{
\link[base]{sort}
}
\keyword{internal}
